+++ openwisp_radius/tests/test_api/test_api.py
@@ -14,6 +14,7 @@ from dj_rest_auth.tests.utils import (
 from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.contrib.auth.models import Permission
+from django.contrib.contenttypes.models import ContentType
 from django.core import mail
 from django.core.cache import cache
 from django.core.mail import EmailMultiAlternatives
@@ -25,6 +26,7 @@ from rest_framework.test import APIClient
 
 from openwisp_radius import settings as app_settings
 from openwisp_radius.api.serializers import (
+    RadiusUserGroupSerializer,
     RadiusUserSerializer,
     UserGroupCheckSerializer,
 )
@@ -38,6 +40,7 @@ User = get_user_model()
 RadiusToken = load_model("RadiusToken")
 RadiusBatch = load_model("RadiusBatch")
 RadiusUserGroup = load_model("RadiusUserGroup")
+RadiusGroup = load_model("RadiusGroup")
 OrganizationRadiusSettings = load_model("OrganizationRadiusSettings")
 Organization = swapper.load_model("openwisp_users", "Organization")
 OrganizationUser = swapper.load_model("openwisp_users", "OrganizationUser")
@@ -1072,6 +1075,486 @@ class TestApi(AcctMixin, ApiTokenMixin, BaseTestCase):
                 "value": "2000000000",
             },
         )
+
+    def _add_model_permission(self, user, model, actions):
+        """action must be one of: 'view', 'add', 'change', 'delete'"""
+        content_type = ContentType.objects.get_for_model(model)
+        actions = actions or []
+        for action in actions:
+            permission = Permission.objects.get(
+                codename=f"{action}_{model._meta.model_name}",
+                content_type=content_type,
+            )
+            user.user_permissions.add(permission)
+
+    def test_radius_group_list(self):
+        org1 = self._create_org(name="Org 1")
+        org2 = self._create_org(name="Org 2")
+        user = self._create_user()
+        self._create_org_user(user=user, organization=org1, is_admin=True)
+        self.assertEqual(RadiusGroup.objects.filter(organization=org1).count(), 2)
+        self.assertEqual(RadiusGroup.objects.filter(organization=org2).count(), 2)
+        url = reverse("radius:radius_group_list")
+
+        with self.subTest("Unauthenticated access"):
+            response = self.client.get(url)
+            self.assertEqual(response.status_code, 401)
+
+        self.client.force_login(user=user)
+        with self.subTest("Access without permission"):
+            response = self.client.get(url)
+            self.assertEqual(response.status_code, 403)
+            response = self.client.post(
+                url, {"name": "Test Group", "organization": org1.id}
+            )
+            self.assertEqual(response.status_code, 403)
+
+        self._add_model_permission(user, RadiusGroup, ["view", "add"])
+
+        with self.subTest("Basic list"):
+            response = self.client.get(url)
+            self.assertEqual(response.status_code, 200)
+            data = response.data
+            self.assertEqual(data["count"], 2)
+            self.assertEqual(data["next"], None)
+            self.assertEqual(data["previous"], None)
+            results = data["results"]
+            self.assertEqual(len(results), 2)
+            result_ids = [item["id"] for item in results]
+            self.assertEqual(
+                RadiusGroup.objects.filter(
+                    id__in=result_ids, organization=org1
+                ).count(),
+                2,
+            )
+
+        with self.subTest("Filtering by organization"):
+            response = self.client.get(url, {"organization": org1.id})
+            self.assertEqual(response.status_code, 200)
+            data = response.data
+            self.assertEqual(data["count"], 2)
+            results = data["results"]
+            self.assertEqual(len(results), 2)
+            result_ids = [item["id"] for item in results]
+            self.assertEqual(RadiusGroup.objects.filter(id__in=result_ids).count(), 2)
+
+        with self.subTest("Search by name"):
+            response = self.client.get(url, {"search": "power-users"})
+            self.assertEqual(response.status_code, 200)
+            data = response.data
+            self.assertEqual(data["count"], 1)
+            results = data["results"]
+            self.assertEqual(results[0]["name"], "org-1-power-users")
+
+        with self.subTest("Search returns no results"):
+            response = self.client.get(url, {"search": "Nonexistent"})
+            self.assertEqual(response.status_code, 200)
+            self.assertEqual(len(response.data["results"]), 0)
+
+        with self.subTest("Create RADIUS group"):
+            response = self.client.post(
+                url,
+                {
+                    "name": "Unlimited",
+                    "organization": org1.id,
+                    "description": "Unlimited description",
+                },
+                content_type="application/json",
+            )
+            self.assertEqual(response.status_code, 201)
+            data = response.data
+            self.assertEqual(data["name"], "org-1-Unlimited")
+            self.assertEqual(data["description"], "Unlimited description")
+
+        with self.subTest("Create RadiusGroup with other organization"):
+            response = self.client.post(
+                url,
+                {
+                    "name": "Test Group Org2",
+                    "organization": org2.id,
+                    "description": "Test group description",
+                },
+                content_type="application/json",
+            )
+            self.assertEqual(response.status_code, 400)
+
+    def test_radius_group_detail(self):
+        org1 = self._create_org(name="Org 1")
+        org2 = self._create_org(name="Org 2")
+        user = self._create_user()
+        self._create_org_user(user=user, organization=org1, is_admin=True)
+        group = RadiusGroup.objects.get(organization=org1, name="org-1-power-users")
+        url = reverse("radius:radius_group_detail", args=[group.pk])
+        with self.subTest("Unauthenticated requests"):
+            # GET operation
+            response = self.client.get(url)
+            self.assertEqual(response.status_code, 401)
+            # PATCH operation
+            response = self.client.patch(
+                url,
+                {"description": "Patched description"},
+                content_type="application/json",
+            )
+            self.assertEqual(response.status_code, 401)
+            # PUT operation
+            response = self.client.put(
+                url,
+                {
+                    "name": "Updated Group",
+                    "organization": org1.id,
+                    "description": "Updated description",
+                },
+                content_type="application/json",
+            )
+            self.assertEqual(response.status_code, 401)
+            # DELETE operation
+            response = self.client.delete(url)
+            self.assertEqual(response.status_code, 401)
+
+        self.client.force_login(user=user)
+        with self.subTest("Operations without permission"):
+            response = self.client.get(url)
+            self.assertEqual(response.status_code, 403)
+
+        self._add_model_permission(user, RadiusGroup, ["view", "change", "delete"])
+
+        with self.subTest("GET operation"):
+            response = self.client.get(url)
+            self.assertEqual(response.status_code, 200)
+            data = response.json()
+            self.assertEqual(data["name"], group.name)
+            self.assertEqual(data["description"], group.description)
+            self.assertEqual(data["id"], str(group.id))
+
+        with self.subTest("GET not found"):
+            fake_uuid = "00000000-0000-0000-0000-000000000000"
+            fake_url = reverse("radius:radius_group_detail", args=[fake_uuid])
+            response = self.client.get(fake_url)
+            self.assertEqual(response.status_code, 404)
+
+        with self.subTest("PATCH operation"):
+            response = self.client.patch(
+                url,
+                {"description": "Patched description"},
+                content_type="application/json",
+            )
+            self.assertEqual(response.status_code, 200)
+            data = response.json()
+            self.assertEqual(data["description"], "Patched description")
+            self.assertEqual(data["name"], group.name)
+
+        with self.subTest("PUT operation"):
+            response = self.client.put(
+                url,
+                {
+                    "name": "Updated Group",
+                    "organization": org1.id,
+                    "description": "Updated description",
+                },
+                content_type="application/json",
+            )
+            self.assertEqual(response.status_code, 200)
+            data = response.data
+            self.assertEqual(data["name"], "org-1-Updated Group")
+            self.assertEqual(data["description"], "Updated description")
+
+        with self.subTest("DELETE operation"):
+            response = self.client.delete(url)
+            self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+            self.assertEqual(RadiusGroup.objects.filter(pk=group.pk).count(), 0)
+
+        with self.subTest("Perform operation on object of other organization"):
+            org2_group = RadiusGroup.objects.get(
+                organization=org2, name="org-2-power-users"
+            )
+            url_org2 = reverse("radius:radius_group_detail", args=[org2_group.pk])
+            # GET operation
+            response = self.client.get(url_org2)
+            self.assertEqual(response.status_code, 404)
+            # PATCH operation
+            response = self.client.patch(
+                url_org2,
+                {"description": "Patched description"},
+                content_type="application/json",
+            )
+            self.assertEqual(response.status_code, 404)
+            # PUT operation
+            response = self.client.put(
+                url_org2,
+                {
+                    "name": "Updated Group",
+                    "organization": org1.id,
+                    "description": "Updated description",
+                },
+                content_type="application/json",
+            )
+            self.assertEqual(response.status_code, 404)
+            # DELETE operation
+            response = self.client.delete(url_org2)
+            self.assertEqual(response.status_code, 404)
+
+    def test_radius_user_group_list(self):
+        org1 = self._create_org(name="Org 1")
+        org2 = self._create_org(name="Org 2")
+        admin_user = self._create_user(username="admin_user", email="admin@test.com")
+        self._create_org_user(user=admin_user, organization=org1, is_admin=True)
+        target_user = self._create_user(username="target_user", email="target@test.com")
+        self._create_org_user(user=target_user, organization=org1)
+        self._create_org_user(user=target_user, organization=org2)
+        # Create a user in org2 that admin_user should not be able to access
+        org2_user = self._create_user(username="org2_user", email="org2@test.com")
+        self._create_org_user(user=org2_user, organization=org2)
+        org1_group = RadiusGroup.objects.get(organization=org1, name="org-1-users")
+        rug = self._create_radius_usergroup(
+            user=target_user, group=org1_group, priority=1
+        )
+        org2_group = RadiusGroup.objects.get(organization=org2, name="org-2-users")
+        self._create_radius_usergroup(user=target_user, group=org2_group, priority=1)
+        url = reverse("radius:radius_user_group_list", args=[target_user.pk])
+        org2_user_url = reverse("radius:radius_user_group_list", args=[org2_user.pk])
+
+        with self.subTest("Unauthenticated access"):
+            response = self.client.get(url)
+            self.assertEqual(response.status_code, status.HTTP_401_UNAUTHORIZED)
+
+        self.client.force_login(user=admin_user)
+        with self.subTest("Access without permission"):
+            response = self.client.get(url)
+            self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+            response = self.client.post(
+                url, {"group": str(org1_group.pk)}, content_type="application/json"
+            )
+            self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+
+        self._add_model_permission(admin_user, RadiusUserGroup, ["view", "add"])
+        with self.subTest("List RadiusUserGroups"):
+            response = self.client.get(url)
+            self.assertEqual(response.status_code, status.HTTP_200_OK)
+            data = response.data
+            self.assertEqual(data["count"], 1)
+            self.assertEqual(data["results"][0]["id"], str(rug.pk))
+
+        with self.subTest("Filter by organization query parameter"):
+            # requesting the same organization that the existing group belongs to
+            response = self.client.get(url + f"?group__organization={org1.pk}")
+            self.assertEqual(response.status_code, status.HTTP_200_OK)
+            self.assertEqual(response.data["count"], 1)
+            # a different organization should yield no results
+            response = self.client.get(url + f"?group__organization={org2.pk}")
+            self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+            self.assertIn("group__organization", response.data)
+
+        with self.subTest("Create RadiusUserGroup"):
+            org1_power_users = RadiusGroup.objects.get(
+                organization=org1, name="org-1-power-users"
+            )
+            response = self.client.post(
+                url,
+                {"group": str(org1_power_users.pk), "priority": 2},
+                content_type="application/json",
+            )
+            self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+            self.assertEqual(
+                RadiusUserGroup.objects.filter(
+                    user=target_user, group__organization=org1
+                ).count(),
+                2,
+            )
+
+        with self.subTest("Cannot access user from other organization"):
+            response = self.client.get(org2_user_url)
+            self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        with self.subTest("Cannot create for user not in admin's organization"):
+            org2_user2 = self._create_user(
+                username="org2_user2", email="org2user2@test.com"
+            )
+            self._create_org_user(user=org2_user2, organization=org2)
+            org2_user2_url = reverse(
+                "radius:radius_user_group_list", args=[org2_user2.pk]
+            )
+            org1_group = RadiusGroup.objects.get(
+                organization=org1, name="org-1-power-users"
+            )
+            response = self.client.post(
+                org2_user2_url,
+                {"group": str(org1_group.pk)},
+                content_type="application/json",
+            )
+            self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        with self.subTest("Cannot create RadiusUserGroup with group from other org"):
+            # target_user is member of org2,
+            # but admin_user can only manage org1
+            org2_group = RadiusGroup.objects.get(
+                organization=org2, name="org-2-power-users"
+            )
+            response = self.client.post(
+                url,
+                {"group": str(org2_group.pk)},
+                content_type="application/json",
+            )
+            self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+            self.assertIn("group", response.data)
+            self.assertEqual(response.data["group"][0].code, "does_not_exist")
+
+            # target_user is only member of org1,
+            # admin_user can manage both org1 and org2
+            OrganizationUser.objects.filter(
+                user=target_user, organization=org2
+            ).delete()
+            self._create_org_user(user=admin_user, organization=org2, is_admin=True)
+            response = self.client.post(
+                url,
+                {"group": str(org2_group.pk)},
+                content_type="application/json",
+            )
+            self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+            self.assertIn("group", response.data)
+            self.assertEqual(response.data["group"][0].code, "does_not_exist")
+
+        with self.subTest("Superuser can access any user"):
+            superuser = self._get_admin()
+            self.client.force_login(user=superuser)
+            response = self.client.get(url)
+            self.assertEqual(response.status_code, status.HTTP_200_OK)
+            response = self.client.get(org2_user_url)
+            self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+    def test_radius_user_group_detail(self):
+        org1 = self._create_org(name="Org 1")
+        org2 = self._create_org(name="Org 2")
+        admin_user = self._create_user(username="admin_user", email="admin@test.com")
+        self._create_org_user(user=admin_user, organization=org1, is_admin=True)
+        target_user = self._create_user(username="target_user", email="target@test.com")
+        self._create_org_user(user=target_user, organization=org1)
+        org1_default_group = RadiusGroup.objects.get(
+            organization=org1, name="org-1-users"
+        )
+        org1_power_users_group = RadiusGroup.objects.get(
+            organization=org1, name="org-1-power-users"
+        )
+        rug = self._create_radius_usergroup(
+            user=target_user, group=org1_default_group, priority=1
+        )
+        url = reverse("radius:radius_user_group_detail", args=[target_user.pk, rug.pk])
+
+        with self.subTest("Unauthenticated access"):
+            response = self.client.get(url)
+            self.assertEqual(response.status_code, status.HTTP_401_UNAUTHORIZED)
+            response = self.client.patch(
+                url, {"priority": 5}, content_type="application/json"
+            )
+            self.assertEqual(response.status_code, status.HTTP_401_UNAUTHORIZED)
+            response = self.client.delete(url)
+            self.assertEqual(response.status_code, status.HTTP_401_UNAUTHORIZED)
+
+        self.client.force_login(user=admin_user)
+        with self.subTest("Access without permission"):
+            response = self.client.get(url)
+            self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+
+        self._add_model_permission(
+            admin_user, RadiusUserGroup, ["view", "change", "delete"]
+        )
+
+        with self.subTest("GET operation"):
+            response = self.client.get(url)
+            self.assertEqual(response.status_code, status.HTTP_200_OK)
+            data = response.json()
+            self.assertEqual(data["id"], str(rug.pk))
+            self.assertEqual(data["group"], str(org1_default_group.pk))
+
+        with self.subTest("PATCH operation"):
+            response = self.client.patch(
+                url, {"priority": 10}, content_type="application/json"
+            )
+            self.assertEqual(response.status_code, status.HTTP_200_OK)
+            rug.refresh_from_db()
+            self.assertEqual(rug.priority, 10)
+
+        with self.subTest("PUT operation"):
+            response = self.client.put(
+                url,
+                {"group": str(org1_power_users_group.pk), "priority": 3},
+                content_type="application/json",
+            )
+            self.assertEqual(response.status_code, status.HTTP_200_OK)
+            rug.refresh_from_db()
+            self.assertEqual(rug.group, org1_power_users_group)
+            self.assertEqual(rug.priority, 3)
+
+        with self.subTest("PUT without group field"):
+            response = self.client.put(
+                url,
+                {"priority": 4},
+                content_type="application/json",
+            )
+            self.assertEqual(response.status_code, status.HTTP_200_OK)
+            rug.refresh_from_db()
+            self.assertEqual(rug.group, org1_power_users_group)
+            self.assertEqual(rug.priority, 4)
+
+        with self.subTest("Org manager cannot assign group from another org"):
+            self._create_org_user(user=target_user, organization=org2)
+            org2_group = RadiusGroup.objects.get(organization=org2, name="org-2-users")
+            response = self.client.put(
+                url,
+                {"group": str(org2_group.pk), "priority": 8},
+                content_type="application/json",
+            )
+            self.assertEqual(
+                response.status_code,
+                status.HTTP_400_BAD_REQUEST,
+            )
+            rug.refresh_from_db()
+            self.assertEqual(rug.group, org1_power_users_group)
+
+        with self.subTest("DELETE operation"):
+            response = self.client.delete(url)
+            self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+            self.assertEqual(RadiusUserGroup.objects.filter(pk=rug.pk).count(), 0)
+
+        with self.subTest("GET not found"):
+            fake_uuid = "00000000-0000-0000-0000-000000000000"
+            fake_url = reverse(
+                "radius:radius_user_group_detail", args=[target_user.pk, fake_uuid]
+            )
+            response = self.client.get(fake_url)
+            self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        with self.subTest("Cannot access RadiusUserGroup of user from other org"):
+            org2_user = self._create_user(username="org2_user", email="org2@test.com")
+            self._create_org_user(user=org2_user, organization=org2)
+            org2_rug = self._create_radius_usergroup(
+                user=org2_user,
+                group=RadiusGroup.objects.get(organization=org2, name="org-2-users"),
+                priority=1,
+            )
+            org2_url = reverse(
+                "radius:radius_user_group_detail", args=[org2_user.pk, org2_rug.pk]
+            )
+            response = self.client.get(org2_url)
+            self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        with self.subTest("Superuser can access any RadiusUserGroup"):
+            superuser = self._get_admin()
+            self.client.force_login(user=superuser)
+            response = self.client.get(org2_url)
+            self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+            # Patch RadiusUserGroup of org2_user
+            response = self.client.patch(
+                org2_url, {"priority": 7}, content_type="application/json"
+            )
+            self.assertEqual(response.status_code, status.HTTP_200_OK)
+            org2_rug.refresh_from_db()
+            self.assertEqual(org2_rug.priority, 7)
+
+    def test_radius_user_group_serializer_without_view_context(self):
+        serializer = RadiusUserGroupSerializer(context={})
+        self.assertEqual(serializer._user, None)
+        self.assertEqual(serializer.fields["group"].queryset.count(), 0)
 
 
 class TestTransactionApi(AcctMixin, ApiTokenMixin, BaseTransactionTestCase):